Refactor: Extract `setupLocalGitRepo` helper for reuse across tests
