+++ internal/cli/install_test.go
@@ -1595,51 +1595,7 @@ func TestPullToStagingDir_RejectsUnsafeArtifactNames(t *testing.T) {
 }
 
 func setupGitRepo(t *testing.T) string {
-	t.Helper()
-	dir := t.TempDir()
-	workDir := filepath.Join(dir, "work")
-	bareDir := filepath.Join(dir, "bare.git")
-
-	run := func(wd string, args ...string) {
-		t.Helper()
-		cmd := exec.Command(args[0], args[1:]...)
-		cmd.Dir = wd
-		cmd.Env = append(os.Environ(),
-			"GIT_AUTHOR_NAME=test", "GIT_AUTHOR_EMAIL=t@t",
-			"GIT_COMMITTER_NAME=test", "GIT_COMMITTER_EMAIL=t@t",
-			"GIT_CONFIG_NOSYSTEM=1", "GIT_CONFIG_GLOBAL="+os.DevNull,
-		)
-		out, err := cmd.CombinedOutput()
-		if err != nil {
-			t.Fatalf("%v failed: %s\n%s", args, err, out)
-		}
-	}
-
-	run(dir, "git", "init", "--bare", "-b", "master", bareDir)
-	run(dir, "git", "clone", bareDir, workDir)
-
-	manifest := `{
-  "apiVersion": "striatum.dev/v1alpha2",
-  "kind": "Skill",
-  "metadata": {"name": "git-skill", "version": "1.0.0"},
-  "spec": {"entrypoint": "SKILL.md", "files": ["SKILL.md"]}
-}`
-	if err := os.WriteFile(filepath.Join(workDir, "artifact.json"), []byte(manifest), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(workDir, "SKILL.md"), []byte("# Git Skill"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	run(workDir, "git", "add", "-A")
-	run(workDir, "git", "commit", "-m", "init")
-	run(workDir, "git", "tag", "v1.0.0")
-	run(workDir, "git", "push", "origin", "HEAD", "--tags")
-
-	p := filepath.ToSlash(bareDir)
-	if len(p) > 0 && p[0] != '/' {
-		p = "/" + p
-	}
-	return "file://" + p
+	return setupLocalGitRepo(t, "", "v1.0.0")
 }
 
 func TestInstall_GitRef_HappyPath(t *testing.T) {
